backend/requirements.txt
Flask==3.0.0
tensorflow==2.15.0
pillow==10.0.0
